Extract bridge HTTP helpers to sgv2-service-common (#1641)
